+++ app/src/routes/editor/EditPanel/inputs/AutoLayoutInput.tsx
@@ -1,4 +1,4 @@
-import { useEditorEngine } from '@/components/Context/Editor';
+import { useEditorEngine } from '@/components/Context';
 import {
     getAutolayoutStyles,
     getInputValues,
