+++ app/src/routes/editor/EditPanel/inputs/TagDetails.tsx
@@ -1,4 +1,4 @@
-import { useEditorEngine } from '@/components/Context/Editor';
+import { useEditorEngine } from '@/components/Context';
 import { TAG_INFO } from '@/lib/editor/styles/tag';
 import { motion } from 'framer-motion';
 import { observer } from 'mobx-react-lite';
