+++ app/src/routes/editor/RightClickMenu/index.tsx
@@ -1,4 +1,4 @@
-import { useEditorEngine } from '@/components/Context/Editor';
+import { useEditorEngine } from '@/components/Context';
 import {
     ContextMenu,
     ContextMenuContent,
