+++ app/src/routes/editor/WebviewArea/ResizeHandles.tsx
@@ -1,8 +1,8 @@
+import { useEditorEngine } from '@/components/Context';
 import { EditorMode } from '@/lib/models';
 import clsx from 'clsx';
 import { observer } from 'mobx-react-lite';
 import { MouseEvent, useRef } from 'react';
-import { useEditorEngine } from '@/components/Context/Editor';
 
 interface ResizeHandleProps {
     webviewRef: React.RefObject<Electron.WebviewTag>;
