+++ app/src/routes/projects/ProjectsTab/Select/index.tsx
@@ -1,4 +1,4 @@
-import { useProjectsManager } from '@/components/Context/Editor';
+import { useProjectsManager } from '@/components/Context';
 import { observer } from 'mobx-react-lite';
 import { useEffect, useState } from 'react';
 import EmblaCarousel from './Carousel';
