+++ app/src/routes/projects/ProjectsTab/index.tsx
@@ -1,4 +1,4 @@
-import { useProjectsManager } from '@/components/Context/Editor';
+import { useProjectsManager } from '@/components/Context';
 import { observer } from 'mobx-react-lite';
 import { CreateMethod } from '..';
 import { ChooseMethod } from './Create/ChooseMethod';
